Refactor authentication middleware to use consistent error handling; update user model save hook and enhance cloud upload error management

+++ src/utils/uploadoncloudinary.js
@@ -1,5 +1,6 @@
 import { v2 as cloudinary } from 'cloudinary';
 import fs from 'fs';
+import errorhandler from './errorhandler.js';
 
 // Configuration
     cloudinary.config({ 
@@ -19,9 +20,11 @@ import fs from 'fs';
             console.log(result.url);
             return result;
         } catch (error) {    
+            if(filePath&&fs.existsSync(filePath)){
             fs.unlinkSync(filePath);
+            }
             console.error("Cloudinary upload error:", error);
-            return null;
+            throw new errorhandler(error?.statusCode || 500,error?.message || "Cloudinary upload failed",[error])
 
         }
     }
